pyproject.toml
[tool.black]
line-length = 80
include = '\.pyi?$'

[project]
name = "forest_gen"
version = "0.3.4"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "opensimplex==0.4.5.1",
    "scipy>=1.0.0",
]
authors = [
    {name = "Tomasz Chady", email = "tomasz.chady@gmail.com"},
    {name = "Jakub Markil", email = "jakub.markil1@gmail.com"},
    {name = "Patryk Olszewski", email = "patryk.olszewski@vip.onet.pl"},
    {name = "Oskar Winiarski", email = "oskarwiniarski09@gmail.com"},
]
license = "GPL-3.0-only"

[tool.setuptools.packages.find]
where = ["."]
include = ["forest_gen", "forest_gen_utils"]

[project.optional-dependencies]
dev = ["black"]
vis = ["matplotlib", "ipykernel"]
export = ["pillow"]

[project.urls]
Repository = "https://github.com/GrafCzterech/forest_gen"
